backend/requirements.txt
langchain
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn
streamlit==1.29.0
requests==2.31.0
googletrans
geopy
faiss-cpu==1.12.0
sentence-transformers==2.2.2
torch
SpeechRecognition==3.10.0
pydub==0.25.1
gTTS==2.4.0
langdetect
deep_translator
pyaudio
pyttsx3
chromadb
sentence-transformers
flask
flask-cors
flask_sqlalchemy
timedelta
werkzeug